remove backcompat for extension titles starting with "WIP:"
Work-in-progress extensions are indicated by having `"wip": true` in their package.json manifest. Previously, in 2018, we considered extensions WIP if their `title` started with `WIP: ` or `[WIP] `. When we introduced the simpler `wip` field on package.json, we left in backcompat for extensions that indicated WIP in their `title`.

It is safe to remove this backcompat because the `wip` field on package.json has been present for ~3 years. (Also, even the `title` field was removed for extensions back in 2018 (in favor of just `name`, like npm).)