better Cache-Control and fixed Vary header for extension registry API responses
The HTTP request from the extension registry listing all the extensions is now cached for 120s (was 30s). This HTTP request needs to fetch ~800kb (compressed) of JSON data from Sourcegraph.com, and caching it for longer will have minimal negative consequences (because extensions don't change very often) and will marginally reduce how often this data needs to be re-fetched and a user needs to wait for it.

The extension registry's Vary header was being set incorrectly. The Vary header should list the names of *request* headers that help determine if a cached response is valid. But we were including *response* headers, which didn't cause any issues (because the requests didn't even include this header).