add `extensionIDs` param to GraphQL extensions queries
This significantly speeds up the `graphql?Extensions` query that needs to run on each page load. Previously, the query to get the user's list of enabled extensions used the `prioritizeExtensionIDs` param (and a `first` param of the length), which means that only their enabled extensions are included in the result. That sounds right, but it doesn't "hint" the intent of the query enough to make it fast. Because the intent is just to sort ("prioritize"), the GraphQL backend still needs to process all possible extensions. (As it turns out, processing them is slow because it must sort a fairly large list based on properties that sometimes require parsing a lot of JSON to determine.)

Using the new `extensionIDs` param (instead of `prioritizeExtensionIDs`) tells the GraphQL backend that it can ignore extensions not listed, so it doesn't need to sort them or otherwise process them.

This speeds up the `graphql?Extensions` timing from ~220ms to ~140ms. It is also a simpler API that more closely matches the intent of this popular way of filtering the extensions results.

Adds tests for GraphQL and DB interactions.